pyproject.toml
[project]
name = "FastAPI-boilerplate"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
